Update Library folders
