+++ WBooks/src/app/screens/Library/screens/index.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import {
+  FlatList,
+  } from 'react-native';
+
+import BookItem from '../components/bookItem';
+import booksList from '../../../../utils/data';
+
+function BookList() {
+    return (
+    <FlatList
+        data={booksList}
+        renderItem={({ item }) => <BookItem title={item.title} />}
+        keyExtractor={item => item.id.toString()}
+    />)
+}
+
+export default BookList;
