+++ WBooks/src/utils/colors.js
@@ -1,0 +1,5 @@
+const colors = {
+    lightBackgroundBlue: '#F2F6FA'
+}
+
+export default colors
